refactor: move mongo-helper to a separated folder
