+++ src/infra/db/mongodb/account/queries/check-account-by-email-mongo-repository.ts
@@ -1,4 +1,4 @@
-import { MongoHelper } from '@/infra/db/mongodb'
+import { MongoHelper } from '@/infra/db/mongodb/helpers'
 import { type CheckAccountByEmailRepository } from '@/application/protocols/account/queries'
 
 export class CheckAccountByEmailMongoRepository implements CheckAccountByEmailRepository {
