+++ tests/infra/db/mongodb/account/account-test-setup.ts
@@ -1,6 +1,6 @@
 import { type Collection } from 'mongodb'
 
-import { MongoHelper } from '@/infra/db/mongodb'
+import { MongoHelper } from '@/infra/db/mongodb/helpers'
 
 const mongoHelper: MongoHelper = MongoHelper.getInstance()
